requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.112.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==2.1.5
openai==1.35.10
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydotenv==0.0.7
python-dotenv==1.0.1
python-multipart==0.0.17
requests==2.32.3
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.32.1
Werkzeug==3.0.3
replicate